+++ fairseq/models/speech_to_speech/__init__.py
@@ -5,5 +5,6 @@
 
 from .modules import *  # noqa
 from .s2s_conformer import *  # noqa
-from .s2s_conformer_t2 import *  # noqa
+from .s2s_conformer_translatotron2 import *  # noqa
+from .s2s_conformer_unity import *  # noqa
 from .s2s_transformer import *  # noqa
