+++ fairseq/models/speech_to_speech/s2s_conformer.py
@@ -11,10 +11,10 @@ import torch
 from fairseq import checkpoint_utils
 from fairseq.models import register_model, register_model_architecture
 from fairseq.models.speech_to_speech.s2s_transformer import (
-    S2SpecTTransformerModel,
     S2UTTransformerModel,
-    s2spect_architecture_base,
+    TranslatotronTransformerModel,
     s2ut_architecture_base,
+    translatotron_architecture_base,
 )
 from fairseq.models.speech_to_text import S2TConformerEncoder
 from fairseq.models.transformer import Linear
@@ -53,7 +53,7 @@ class S2SConformerEncoder(S2TConformerEncoder):
 @register_model("s2ut_conformer")
 class S2UTConformerModel(S2UTTransformerModel):
     """
-    Direct speech-to-speech translation model with S2T Conformer encoder + Transformer discrete unit decoder
+    Direct speech-to-speech translation model with Conformer encoder + Transformer discrete unit decoder (S2UT)
     """
 
     @staticmethod
@@ -90,15 +90,15 @@ class S2UTConformerModel(S2UTTransformerModel):
         return encoder
 
 
-@register_model("s2spect_conformer")
-class S2SpecTConformerModel(S2SpecTTransformerModel):
+@register_model("translatotron_conformer")
+class TranslatotronConformerModel(TranslatotronTransformerModel):
     """
-    Direct speech-to-speech translation model with S2T Conformer encoder + TTS Transformer decoder
+    Direct speech-to-speech translation model with Conformer encoder + TTS Transformer decoder (Translatotron)
     """
 
     @staticmethod
     def add_args(parser):
-        S2SpecTTransformerModel.add_args(parser)
+        TranslatotronTransformerModel.add_args(parser)
         parser.add_argument("--depthwise-conv-kernel-size", type=int, default=31)
         parser.add_argument(
             "--attn-type",
@@ -146,8 +146,8 @@ def s2ut_conformer_architecture_base(args):
     s2ut_architecture_base(args)
 
 
-@register_model_architecture("s2spect_conformer", "s2spect_conformer")
-def s2spect_conformer_architecture_base(args):
+@register_model_architecture("translatotron_conformer", "translatotron_conformer")
+def translatotron_conformer_architecture_base(args):
     args.attn_type = getattr(args, "attn_type", None)
     args.pos_enc_type = getattr(args, "pos_enc_type", "abs")
     args.input_feat_per_channel = getattr(args, "input_feat_per_channel", 80)
@@ -159,11 +159,13 @@ def s2spect_conformer_architecture_base(args):
     args.dropout = getattr(args, "dropout", 0.1)
     args.encoder_layers = getattr(args, "encoder_layers", 16)
     args.depthwise_conv_kernel_size = getattr(args, "depthwise_conv_kernel_size", 31)
-    s2spect_architecture_base(args)
+    translatotron_architecture_base(args)
 
 
-@register_model_architecture("s2spect_conformer", "s2spect_conformer_fisher")
-def s2spect_architecture_fisher(args):
+@register_model_architecture(
+    "translatotron_conformer", "translatotron_conformer_fisher"
+)
+def translatotron_architecture_fisher(args):
     args.encoder_embed_dim = getattr(args, "encoder_embed_dim", 256)
     args.encoder_ffn_embed_dim = getattr(args, "encoder_ffn_embed_dim", 256 * 8)
     args.encoder_attention_heads = getattr(args, "encoder_attention_heads", 4)
@@ -172,4 +174,18 @@ def s2spect_architecture_fisher(args):
     # decoder
     args.prenet_dim = getattr(args, "prenet_dim", 32)
 
-    s2spect_conformer_architecture_base(args)
+    translatotron_conformer_architecture_base(args)
+
+
+# for old models
+@register_model_architecture(
+    model_name="translatotron_conformer", arch_name="s2spect_conformer"
+)
+def translatotron_conformer_architecture_base_legacy(args):
+    translatotron_conformer_architecture_base(args)
+
+
+# for old models
+@register_model_architecture("translatotron_conformer", "s2spect_conformer_fisher")
+def translatotron_architecture_fisher_legacy(args):
+    translatotron_architecture_fisher(args)
